requirements.txt
asgiref==3.7.2
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
sqlparse==0.4.4
